Tidied up the build process somewhat, especially for HECToR.
Both libgc and libgmp are now linked statically rather than dynamically.
That not only prevents unnecessary linking of extra libraries on HECToR,
but it is also probably better for other systems (since the libraries
are not actually shared with other applications).

It is now possible to specify the C compiler as command line argument.
The C compiler setting is also properly propagated to all subprojects
under all circumstances.

There is now a message for which subproject is being built.

Cleaned up the Makefile to have less repetition.

Standardize the build process around ZeroMQ 3.2.0, so that we don't
have different versions on different systems (even if the wire formats
are compatible).