System environment variables can now be accessed via OS_ENV.
OS_ENV is an immutable record of strings, containing a binding from an
environment variable name to its associated value. This is to support
querying of environment variables, allowing both the use of standard
environment variables such as TERM and the use of GAP-specific
environment variables (such as the IP/port of the master process in a
distributed environment).