Added automatic readlocking functionality to the kernel.
Regions can now be declared as being automatically readlocked with the
SetAutoLockRegion(obj, bool) function. Such a region will automatically
be readlocked upon access if necessary. This lock will automatically be
released when one of the following happens:

* The surrounding GAP function returns.
* A surrounding atomic statement ends.
* At the end of a command entered at the command line.
* At the end of a file being read.

There are convenience functions ShareAutoReadObj() and
NewAutoReadRegion() with the same semantics as ShareObj() and
NewRegion(), except that they set the autolock flag on the resulting
region.