Temporary fix for SET_FILT_LIST() breaking.
SET_FILT_LIST() is frequently being called on objects to which the code
has only read access, causing a write guard error when the type of the
object is being changed.

This temporary fix avoids retyping when the bag is not writable. Because
code generally does not re-inspect the bag to check the tnum again (and
should not, even once we have a more comprehensive solution), this
should be safe for the time being.