8bit MultiplicativeElementsWithInversesFamilyByRws now uses ConstructExtendedType.
This mimics the current state of the 16bit method and fixes a bug in grppcnrm.tst.
Still this is puzziling since there shouldn't be a difference in behavior (other
than how regions are managed, which is not at issue), because BindOnce() in this
context (assigning to unbound entries in a positional object) should behave
exactly like an assignment. We should investigate this.