Added functionality for object sets and maps.
Object sets and maps are created by OBJ_SET() and OBJ_MAP()
respectively. They are hash tables that use object identity for
comparison. They can be manipulated with the following functions:

* set := OBJ_SET();
* set := OBJ_SET(list_of_initial_values);
* ADD_OBJ_SET(set, obj);
* REMOVE_OBJ_SET(set, obj);
* bool := FIND_OBJ_SET(set, obj);
* CLEAR_OBJ_SET(set);

* map := OBJ_MAP();
* map := OBJ_MAP(list_of_initial_values);
* ADD_OBJ_MAP(map, key, value);
* REMOVE_OBJ_MAP(map, key);
* bool := FIND_OBJ_MAP(map, key, default_value);
* CLEAR_OBJ_MAP(map);

For OBJ_MAP(), the list of initial values must have a length
that is a multiple of two and is treated as a flattened list
of key/value pairs.