Implemented WriteLock() and ReadLock() functions.
WriteLock() and ReadLock() take an arbitrary number of shared
objects and lock the corresponding regions. They must be used
from within an atomic function or section; if they're used
outside of those constructs, an error will be raised.

The intended use is to avoid an additional level of indentation in
functions that use locks. Instead of:

    f := function()
        atomic R do
	    # code ...
	od;
    end;

one can do:

    f := atomic function()
        WriteLock(R);
	#code ...
    end;