Made T_DATOBJ objects readonly when they are being made immutable.
Most of the time, data objects that are being made immutable in
the mathematical sense actually have a fixed representation. This
branch enforces this discipline by moving them to the read-only
region, where further modifications will result in an error.

If modifications of an immutable object is needed, it should be
made public by the corresponding code and operations should be
atomic (e.g., by using the hash lock functionality).