Fixed Add() behavior for atomic lists.
Previously, the behavior of ADD_LIST() was not atomic for atomic
lists. It calculated the length of the list, then assigned an element
to the position past that. However, the size of the list could change
in the meantime.

This is now an atomic operation; there is also now a new function
AddAtomicList(), which behaves like Add(), but also returns the new
length of the list.