Changes to make RunTask(QuaternionAlgebra,Rationals) working.
I observed that it is not working being called several times as a task, since some data went to the thread-local region of the first thread, and then second time another thread tried to access them and failed. The main change is that IsDenseCoeffVectorRep is made IsAtomicPositionalObjectRep. Also, QuaternionAlgebraDatastores immutable lists now.