Changed DISABLE_GUARDS functionality and implementation.
DISABLE_GUARDS() is now a function that takes a boolean or integer
argument. The meaning of the argument is as follows:

* 0 or false: Guards work normally.
* 1 or true: Guards are disabled and will not produce errors.
* >= 2: Guards are disabled and explicit checks that try to determine
  whether a thread has read or write access to a region will also
  always return true.

Finally, the effect of DISABLE_GUARDS() is now thread-local. While it is
still not safe to use in a production environment, this makes it
somewhat less unsafe when debugging.