First stage of dynamic retyping changes.
The dynamic retyping of plain lists has been reworked so that it
respects read- and write-access constraints properly. Lists to which a
thread has write access will have their type updated dynamically as
before; lists to which a thread has only read access will calculate
a more specific type on the fly, if at all. If a thread does not have
access to elements in a list, the types are still determined as
accurately as possible.

To mitigate the need of having to redo calculations for lists to
which a thread does not have read access, sharing and migrating
thread-local data will calculate the most specific type beforehand;
similarly when a list is being made read-only.