Made MakeImmutable() safe.
MakeImmutable() could potentially leave an object with subobjects in
different regions in an undefined state; to avoid this, there is now
a function that verifies that all subobjects are either writable or
already immutable and which gets called prior to making an object
immutable. To access this functionality, call CheckedMakeImmutable()
instead of MakeImmutable() in the kernel; the kernel is still allowed
to call MakeImmutable() and avoid the overhead of the check if it
knows that the call is safe.