Fixed a bug related to the use of GC_size() in ResizeBag().
In some cases, the Boehm GC stores additional information in
the last word of a memory block. GC_size() erroneously includes
the space for that last word in its result; thus, ResizeBag()
could occasionally overwrite the information contained in the
last word if it relied on additional memory being there.

This is fixed for now by not relying on GC_size(), at the cost
of it being slightly less efficient in some rare cases.