Added TICKER() function.
TICKER(count, func) is a fast way to execute a function every
'count' occurrences. It is implemented in C in the kernel; an
equivalent GAP implementation would be along the following
lines:

TICKER_COUNT := 0;
MakeThreadLocal("TICKER_COUNT");

TICKER := function(count, func)
  TICKER_COUNT := TICKER_COUNT + 1;
  if TICKER_COUNT >= count then
    TICKER_COUNT := 0;
    func();
  fi;
end;

The intended usage is in conjunction with OnTaskCancellation() in
inner loops to avoid the overhead of frequent task state checks.