Introduced proper signal handling for threads.
In order to be compatible with proper pthread signal handling, all
signals are now handled in a separate thread, following roughly
the skeleton described in the standard [1].

The new GAP functions used to implement this are SIGWAIT() and
DEFAULT_SIGXXX_HANDLER() (where XXX is one of INT, CHLD, VTALRM, and
WINCH at the moment).

SIGWAIT() accepts a single argument, which is a record that maps
signal names to signal handlers (signal handlers are parameterless
functions). A straightforward implementation is:

CreateThread(function()
  local handlers;
  handlers := rec(
    SIGINT := DEFAULT_SIGINT_HANDLER,
    SIGCHLD := DEFAULT_SIGCHLD_HANDLER,
    SIGVTALRM := DEFAULT_SIGVTALRM_HANDLER,
    SIGWINCH := DEFAULT_SIGWINCH_HANDLER
  );
  while true do
    SIGWAIT(handlers);
  od;
end);

Also, TICKER has been renamed to PERIODIC_CHECK() and the first
argument is now the minimum number of milliseconds between invocations
of the second argument. In order for this to work,
DEFAULT_SIGVTALRM_HANDLER has to be periodically called by SIGWAIT().

Thus, if one wishes to install a new timer handler, the original handler
still has to be invoked as part of it.

[1] http://pubs.opengroup.org/onlinepubs/009695399/functions/pthread_sigmask.html