Tasks now use GAPInfo.KernelInfo.NUM_CPUS worker threads.
On hyper-threaded architectures, this is actually 2x the
number of cores, even though it is unlikely that we can
actually run that many threads in parallel.

The value can be overridden with the -P commandline option.

scons also accepts a cpus=... option to change it during the
build process. Use: "make SCONSFLAGS=cpus=4", for example,
to get a predefined value of 4 for the number of processors.