Added support for deadlock detection.
Regions now have a numeric precedence attribute and need to be
locked in order of descending precedence, unless all locks are
simultaneously acquired as part of a single atomic statement
or LOCK() call.

This is currently disabled by default and has to be enabled with
the -Z command line option.