Fixed deadlock in UpdateCopyFopyInfo().
UpdateCopyFopyInfo() acquires the global gvar lock, but so do its
internal calls to GVarName(). If the lock implementation is not
recursive, this results in undefined behavior, usually a deadlock.

We are now tracking explicitly if the lock is already acquired by
the current thread, and if so, we increment a counter rather
than re-acquiring it. Note that this would not allow for recursive
read locks, but the code does not use read locks recursively
(except when a write lock has already been acquired).