Added support for internal vs. outwardly visible mutability.
Previously, GAP did not distinguish between the abstract and the
concrete state of an object when it came to mutability. Mutability
was considered solely a property of the abstract, outwardly visible
state. This change introduces a separate concept, mutability of the
concrete state. This is implemented through a separate category,
IsInternallyMutable. This is currently used only for data objects,
where absence of the filter indicates to the kernel that it is
allowed to migrate the object to the read-only region when it is
made immutable. The filter is not set by default.

In the future, this filter may also be useful in conjunction with
Objectify() to use a proper representation/region directly for data
objects, positional objects, and component objects.

Todo: There is currently only an implementation for data objects,
returning false for all other objects. The implementation should be
extended to all objects once the semantics can be considered stable.