Added basic support for actors.
Actors, unlike tasks, are autonomous objects that receive messages
and continue to keep their state between message invocations. This
makes them a useful alternative paradigm to pure task parallelism,
especially in distributed systems, where sharing state between
interdependent tasks can be difficult.

An actor is started with the StartActor(handlers) function. It accepts
a record of functions that describe the messages it accepts (as well
as special functions). It returns a new record with the same field
names, but with functions that operate asynchronously and in the context
of the actor thread rather than synchronously in the current thread.

Note that each running actor occupies a worker thread until it
terminates; this worker thread will not be available run tasks. This
means that there need to be enough worker threads to allow for all
actors to run; also, idle actors cut down on the potential parallelism.
If idle actors are a problem, a possible solution is to use the -P
command line option to increase the number of worker threads beyond the
number of actual physical cores available.

Actors can start new tasks (or even new actors) in order to initiate
new computations or send messages to other actors.

Example (demo/actortest2.g):

  MakeThreadLocal("MyCounter");

  actor := StartActor( rec(
    _Init := function() MyCounter := 0; end,
    _AtExit := function() Display(MyCounter); Unbind(MyCounter); end,
    Inc := function() MyCounter := MyCounter + 1; end,
    Dec := function() MyCounter := MyCounter - 1; end,
    Finish := function() ExitActor(); end
  ) );

  # The following calls are asynchronous rather than synchronous
  # and return immediately; they will be executed when the actor
  # receives the underlying message.

  actor.Inc();
  actor.Inc();
  actor.Dec();
  actor.Finish();

_Init() and _AtExit() are special functions that are called when the
actor is initialized and terminates, respectively. Inc() and Dec() are
the primary messages that the actor responds to. Finish() is used to
terminate the actor (and free up the worker thread). It uses the
library function ExitActor(); ExitActor() will terminate the actor
immediately, discarding any remaining incoming messages.

Note that MyCounter is a thread-local variable so that multiple
instances of the same actor can run concurrently if desired.