Initial serialization support.
Currently supports serialization to and from strings in a
non-portable native format. Use:

	str := SerializeToNativeString(obj);
	obj := DeserializeNativeString(str);

InstallSerializer() and InstallDeserializer() from lib/serialize.g allow
the portable serialization and deserialization of data, positional, and
component objects.

InstallSerializer() takes a description, a filter, and a
function as its arguments. It installs a method
SerializableRepresentation() with those characteristics; this
method should return a list of objects, whose first argument is
a string descriptor.

InstallDeserializer() takes a string descriptor and a function as
its arguments. The string descriptor has to match that returned
by the serializer above; the function must take one less argument
than the number of elements in the list returned by the serializer
above. Upon deserialization, the function is called with all elements
in the list (except the string descriptor) as its arguments. It must
return the new object.