Avoid out-of-memory errors as a result of bad deserialzation input.
Previously, it was possible to construct a serialization string
with huge length fields exceeding the available memory; this would
result in out-of-memory errors when the deserializer attempted to
allocate an object to hold that much memory.

Now, we check that the length fields do not actually exceed the
length of the serialized string.