Added SyncIsBound() and SyncTryWrite() functions.
SyncIsBound() tests if a synchronization variable has been written to
and returns true if that is the case, false otherwise.

SyncTryWrite() attempts to assign a value to a synchronization variable;
if the synchronization variable has already been initialized, it returns
false; otherwise, it performs the same actions as SyncWrite() and returns
true.