Improved the build process.
Fixed some issues with the build process where SCons got confused
about the directories where files were put if the default gcc
wasn't used. Added additional tweaks to facilitate compilation
with gcc-4.8 and to use ccache. To use ccache as a compiler,
create a file with the following two lines:

#!/bin/sh
ccache gcc "$@"

Make it executable and pass the path to it as the COMPILER=
option to make or the compiler= option to SCons.