Added define -DHPCGAP and GAP variable HPCGAP.
This define will be used for all source files as well as all packages
that use bin/cflags; it allows #ifdef HPCGAP/#ifndef HPCGAP preprocessor
directives to distinguish between HPCGAP and GAP.dev, especially when
compiling packages that are meant to work for both.

Similarly, there is now a global variable HPCGAP that can be tested for
with IsBound(HPCGAP). While technically IsBound(CreateThread) and
similar tests would work also, the new variable allows intent to be
documented more clearly and will also keep working if/when we add stub
implementations of concurrency primitives to GAP.dev.