Fixed mutability of the result returned by MakeShiftedVecs().
MakeShiftedVecs() created (and returned) an immutable list containing
mutable data objects. GAP assumes that all subobjects of an immutable
list are also immutable, which fails here and creates problems with
concurrent accessibility (because subobjects remain thread-local to
the thread that originally created them).

To fix this issue, we instead make the list and its subobjects
read-only.