Made TLS allocation more robust against compiler optimizations.
The main thread's TLS allocation scheme needs to manipulate the stack
to arrive at a proper 2^k boundary. We make the relevant variables
volatile to ensure that they are not being optimized away and also
make the dummy variable size 1 to ensure that references to it are
not converted into a null pointer.