Do not compile precompiled modules.
Precompilation currently does not work, because gac does not understand
atomic statements or functions, producing incorrect code. Thus, we do
not need to compile the precompiled modules for the time being; even
when they are compiled, they are not loaded, because they are out of
sync with their original .g files.