Fixed bug in ConvGF2Vec() introduced with SetTypeDatObj() use.
When using SetTypeDatObj(), which could make an object immutable,
the RetypeBag() had to be moved before that statement. However,
it also moved it before a HAS_FILT_LIST() test, causing it to
always fail. The RetypeBag() call is now in its proper place
where it doesn't interfere with that test.