Fixed bad test for AUTO variables.
When an AUTO gvar is first accessed, its stored value is checked to
be non-null; if it is null (unbound) the function is called and the
value is checked again. The second test, however, failed to re-read
the value from gvar storage; thus, it always failed the first time
directly after initialization.

The value is now properly re-read before being tested again.