Make sure that name for SHARE(obj, name, priority) is accessible.
The underlying bug occurred when in SHARE() and SHARE_RECURSE() the
name was also contained in the object itself, e.g.:

  name := "some region";
  r := rec(tag := name);
  ShareObj(r, name);

In this case, sharing the object also migrated the name object to
the shared region, which could then no longer be accessed. This is
now addressed by making a copy of the name first.