Optimized thread-local allocation for Boehm GC.
The build process now supports applying patches to external
libraries. We're using this to apply modifications to the Boehm
GC to make thread-local allocations more efficient.

These modifications are:

* GC default blocksize is set to 64k.
* GC thread-local free lists work for objects up to 256 words (i.e.,
  1024 bytes on a 32-bit architecture, 2048 bytes on a 64-bit
  architecture).

The above numbers can be further modified if needed, though a
performance decrease can be expected if the maximum size of a
thread-local allocation is made too large.

To disable these modifications, build with "scons gc=bdw" (the
default setting is "gc=bdw-tl") or with "make SCONSFLAGS=gc=bdw".

Note that a "make distclean" may be required beforehand to actually
rebuild the garbage collector if this isn't a fresh build.