Added execution framework for ZeroMQ.
To use, run bin/zgap -N <n>. This will create a master node with
an interactive shell and <n> worker nodes that can be controlled
by the master node. Further arguments can be passed on the command
line, e.g.:

	bin/zgap -N 4 -P 8 common.g

Worker nodes are numbered 1 through <n>. The following functions
are available in each worker and master node:

ZBind(node, varname, value);
ZUnbind(node, varname);
ZCall(node, func, args);
ZQuery(node, func, args, callback);
ZTask(node, func, args, callback);
ZAsync(node, func, args);
ZRead(node, file);
ZReadGapRoot(node, file);
ZEval(node, str, callback);
ZExec(node, str);

Callbacks are function that take one argument and that will be
called whenever the executed code is finished. Callbacks that
allow for idle waiting can be constructed with ZResponse().
ZResponse() is a wrapper around a synchronization variable;
it returns a record with 'put', 'get', and 'test' methods.
The 'put' method can be passed as a callback, the result can
be retrieved with 'get' when ready, and 'test' can be used
to test for the availability of the result.

Also, it is possible to execute code on multiple nodes, e.g.:

	ZBind(1, "StartVal", 1);
	ZBind([2..4], "StartVal", 0);

This will set StartVal to 1 on node 1 and to 0 on nodes 2..4. Both
lists and ranges can be used as node descriptors.

The special value ZAll contains a list of all worker nodes. The
master node has id 0, and worker nodes can use the above functions
to execute code on the master also.