Made serialization of typed objects more efficient.
The serialization of typed objects where only the type needs to be
updated can now be done by having an integer or string tag represent the
type in the serialized version. This creates a speedup over the more
general method, which requires a call to the method
"SerializableRepresentation()".

Also, this patch introduces serialization tags for compressed GF(p^k)
vectors and matrices, so that these can be serialized without first
implementing serialization code.