Updated thread-local GC patch to be safe with future versions.
Future versions of the Boehm GC expect mapped memory to begin at an
address that is a multiple of the HBLKSIZE constant. In our patch, we
artificially increase HBLKSIZE (because this determines the size of a
thread-local freelist that we can request). For this to be safe, we call
mmap() with an increased size parameter, then unmap the part at the
beginning and end so that the mapped area begins at a multiple of
HBLKSIZE.