Upgraded Boehm GC and fixed issues with thread-local allocation.
The new GC version is the one of July 21, 2013.

The thread-local allocation patch has been refined to move mark stacks
to static memory instead of keeping them on the stack. Mark stacks can
be very large and can potentially overflow on some architectures.

This also fixes an mmap() alignment error with newer versions of the
GC in conjunction with the patch and a build regression due to which
the GC was compiled without optimization.