Added option for shared region to bypass deadlock detection.
It is now possible to give shared regions a negative precedence value for the
purpose of deadlock detection. This means that such regions will be skipped by
the deadlock detection algorithm. It is the responsibility of the programmer
to use them only in such a way that deadlock cannot result.

Such regions can be created with the NewSpecialRegion(), ShareSpecialObj() and
ShareSingleSpecialObj() functions.