Added more configuration options to the Boehm GC.
It is now possible to build with "make GCBLKSIZE=<size>
GCMAXTHREADS=<num>" (either parameter is optional).

GCBLKSIZE describes the size of individual blocks that the GC manages;
it can be one of: auto, 4096, 8192, 16384, 32768. This value is only
used when GC=boehm-tl or GC=boehm-par is set also and is ignored for
GC=boehm (boehm-tl is the default). The blocksize, in particular,
determines the size of the thread-local allocation pools uses; a larger
value means that thread-local pools need to be refilled less frequently
by a global allocation (which requires a shared lock) but also requires
more thread-local pools (there are a few hundred of these per thread to
account for different allocation sizes and different memory layouts).

GCMAXTHREADS determines the maximum number of worker threads that the
system will support (and also the default number of worker threads).
Specifying a high number of worker threads will not necessarily lead to
better performance, especially on NUMA architectures with highly
asymmetric memory access costs. In particular, a high number of GC
threads can defeat the benefits of a high -m value and vice versa,
because more remote memory accesses are needed.

The actual number of GC worker threads can be set at runtime with the -G
option; if the option is not given, it defaults to the value provided
with -P, and if neither is given, to GCMAXTHREADS.

Both -G and the value of GCMAXTHREADS will be silently ignored if the
system is not built with GC=boehm-par, and thus uses single-threaded
mark-sweep for garbage collection.