Temporarily switch tasks implementation for alpha release.
The reference task implementation currently implements a couple of
features that aren't available in lib/tasks.g (in particular,
ScheduleTask()). We're therefore temporarily switching over the
default to use the reference implementation in order to have those
features in the alpha release. Setting the environment variable
GAP_WORKSTEALING will revert to the work-stealing implementation.