Modified ADJUST_FIELDS_VEC8BIT(v,w) to return 'fail' or [v1,w1]
Formerly it changed v and w in place. Now it calls CopyToVectorRep and creates new compressed vectors equal to v and w, which then are returned in a list of length 2. ADJUST_FIELDS_VEC8BIT is not used in currently redistributed packages, and all its occurences in the GAP library were adjusted to the new behaviour. One should be aware of the situations when ADJUST_FIELDS_VEC8BIT may be called repeatedly to convert the same vector - in this case something should be done elsewhere to fix such inefficiency.