Introduces support for objects that cannot safely be migrated.
Using external libraries can be a problem in that objects created
by such a library may not be threadsafe. We therefore introduce
the concept of fixing an object's region. Objects with a fixed
region cannot be migrated or shared; generally, they will stay within
the thread-local region of the thread that created them for their
entire lifespan. An object's region can be fixed either in the kernel
or in GAP code. In the kernel, use:

    SET_OBJ_FLAG(obj, FIXED_REGION);

In GAP code, use:

    FIX_OBJ_REGION(obj);

After such a call has been performed, MigrateObj(), ShareObj(), etc.
will fail with an error message if they are used to migrate such an
object.