Account for unbound values in thread-local records.
ThreadVar can contain unbound values for thread-local variables
declared by the kernel that have not been initialized. This happens
specifically for the "~" pseudo variable. When printing records
with unbound variables, this would result in a crash, so this
commit fixes printing of thread-local records to handle unound
entries by printing <undefined> for them.