Added a function CompareAndSwapObj().
The new function is defined in <aobjects.h> and implements the usual
compare-and-swap functionality for GAP objects. The first argument
is the address of the object to be swapped out, the second the old
value, and the third the new value. It returns a non-zero value if
the compare-and-swap operation succeeded, zero otherwise.