Tidied up locking in the core library.
Various files used LOCK() instead of alternatives which raise errors
when a lock cannot be acquired, leading to cryptic and delayed error
messages. DO_LOCK() is like lock, but raises an error. WRITE_LOCK()
and READ_LOCK() have also been altered to raise an error when they
cannot acquire a lock (usually due to deadlock constraints).

In addition, filter1.g, oper1.g, and type1.g have been changed to
no longer use the atomic statement but to use the above locking
primitives instead. This should (at least in theory) allow them
to be precompiled again, which is a performance concern.