Make WITH_TARGET_REGION() more robust.
Ideally, we want to have MAKE_TARGET_REGION() as a regular primitive,
not just something that has to be used with care. The primary issue
is that error handling has to revert the current region to the normal
thread-local region (so that objects allocated during error handling
do not end up in the shared target region).

We also fix printing so that any objects allocated because of output
(Print(), Display() in particular) during WITH_TARGET_REGION() is moved
to the public region to avoid access errors.

There may still be further instances of bags being allocated as a
result of internal operations that do not automatically end up in the
public or thread-local region. Further review may be necessary to
identify such objects.