Performance/correctness fix in TakeInterrupt().
HPCGAP's signal handling has changed so that signals are handled
by a separate thread. However, TakeInterrupt() still checked the
global variable syLastIntr in order to look for interrupts. This
was both a correctness and a performance problem.

With respect to correctness, SetInterrupt() now handles interrupts
by remapping the thread-local execution table and is called from
a separate thread that handles interrupts (or manually via !break
from the console UI); this is the state that needs to be checked for.

Performance-wise, the frequent updates to syLastIntr in the
function SyIsIntr() led to a performance degradation, especially
on NUMA architectures. This change leaves the code in place, since
other code may still depend on the function's existence, but in
the long term, remaining usage of SyIsIntr() needs to be eliminated.