Fix speed regression in pc collectors
We were looking up the collector stacks by quering GAP variables in tight
loops in the C code using VAL_GVAR. Instead of that, we now use InitGlobalBag
to setup C variables to directly reference those stack objects.

As an added side benefit, those stacks are now hidden from the user, and
can thus not be messed with. This also means that the user cannot
manually shrink them. To compensate, a new function SET_SCOBJ_MAX_STACK_SIZE
was provided to allow a force-reset of the stacks.