zmodnz: consistently access the modulus
Previously, we stored the modulus in three places:
1. in the type object, accessed via DataType( TypeObj( x ) ),
2. in family object, accessed as fam!.modulus,
3. in family object, accessed as  Characteristic(fam).

To retrieve the modulus for a ZmodnZObj, one can moreover also use
ModulusOfZmodnZObj(x) or Characteristic(x).

This patch changes the code to consistently use Characteristic(fam),
and gets rid of fam!.modulus. The DataType is not use anymore, but is
retained for now, as at least the Browse package is using it. Once
that is changed, we can also get rid of the corresponding three
SetDataType calls in ffee.gi and zmodnz.gi.