Make FirstOp, NumberOp, ForAllOp, ForAnyOp, FilteredOp obsolete
For each Foo in [ First, Number, ForAll, ForAny, Filtered ], there used
to be both Foo (a global function) and FooOp (an operation). Apparently
this was meant as an optimization. However, benchmarking reveals that
this yielded no measurable advantages. Worse, in some cases the default
implementations in the global function were slower than those invoked by
method dispatch.

Now there is just Foo (an operation), and FooOp is a synonym for it,
for compatibility with the few packages installing methods for any of
these (cvec, fr, resclasses).