Add comment to special methods for empty lists
There are special methods for Filtered, ForAll and ForAny which are
applied to empty lists. These provide a tiny but measurable speedup,
so add a comment explaining that (and remove an older comment which
asked about this).

Also, for ForAny and ForAll, refine the selector of the first argument
so that we can use InstallMethod instead of InstallOtherMethod.