src: remove call to mp_set_memory_functions
In Nov 2010, we added code which set GMP to use custom code for memory
allocations, which made it use mmap/munmap. However, we do not need that at
all, as we exclusively use low level GMP interfaces, which do not make any
memory allocations.

Presumably, this change was done to allow us to call high-level GMP APIs, but
we never ended up actually doing that.

Despite this, the change has a noticeable impact, but sadly a negative one: If
one actually does use the high-level APIs, operations become much, much
slower. For example, running Normaliz via NormalizInterface from inside GAP, a
computation that usually takes a dozen seconds or so, did not finish after
several minutes.

If we want to do something like this, I fear we need to implement a proper
full-blown memory manager. But for now, I don't see a reason to do so. Esp. on
64bit machines, were it (empirically) seems relatively safe to use malloc
inside GAP kernel code...