Speeding up loops over iterators
In loops over iterators which are not lists the method selection for
IsDoneIterator and NextIterator can lead to considerable overhead.

This patch introduces a filter `IsStandardIterator` which means that
the iterator is a component object with components .IsDoneIterator
and .NextIterator which provide the methods with the same name.

In the kernel we use these functions directly instead of the generic
operations for iterators in `IsStandardIterator`.

Documentation for `IsStandardIterator` is included. All iterators
constructed by `IteratorByFunctions` are in `IsStandardIterator`.

Examples: Looping over GF(1000003) becomes 25% faster, looping over
SymmetricGroup(11) about 15% faster on my system. Of course, the
advantage is bigger if NextIterator is cheaper.