Prevent inappropriate removal of generators
Prevent the inappropriate removal of the One of a collection of
generators for a semigroup, monoid, inverse semigroup or inverse monoid
in the functions SemigrouByGenerators, MonoidByGenerators,
InverseSemigrouByGenerators, and InverseMonoidByGenerators.

Previously, if the One of a collection <coll> of generators was present
in <coll>, and we tried to make a semigroup/monoid/inverse semigroup or
inverse monoid generated by <coll>, then the One was removed. In the
case that the One was the only generator, or that the One of <coll> was
not the same as the One of <coll> with One removed (possible for partial
perms), this meant that the value of GeneratorsOfMonoid could not be
used to recreate the object. For example,
Semigroup(IdentityTransformation) formerly resulted in a monoid with 0
generators, but Monoid([]) returned an error (it is not possible to
create monoids with 0 generators). This change fixes this, and adds some
tests.