Replace and consolidate ViewString methods for semigroups
Replace the ViewString methods for transformation and partial perm
semigroups by a single function, introduce a new function for semigroups
that happen to be groups.

This changes the ViewObj of transformation and partial perm slightly to
remove "pts" and replaces it with degree/rank for better readability, the
previous methods is consolidated into a single function, which applies
to all semigroups, and the old code is removed.

Update the tests and documentation, removing old view string methods,
new view string method for symmetric inverse monoid, new method for
RankOfPartialPermSemigroup for a partial perm semigroup with generators
of a *group* (this is used by the new ViewString method).