Fix caching of the degree of transformation
This commit fixes the caching of the degree of a transformation.
Previously, although there was space reserved in a transformation for
caching the degree, it was in fact never cached. This caused a problem
in HASH_FUNC_FOR_TRANS where it was assumed that the degree was cached
after it was calculated. In particular, 4 bytes transformations were
trimmed using the (non-)cached value of the degree, which was 0, and
this resulted in non-identity transformations having their value changed
to the identity transformation when they were hashed.