Add method for Socle for finite nilpotent groups
This commit adds a method of computing the Socle for finite nilpotent groups,
computing it as the Omega-parts of the Center.
This method is ranked as high as if it were for only finite nilpotent groups,
Therefore this method will run even if the group does not yet know whether
it is a nilpotent group or not. This method forces an IsNilpotent check after
a CanComputeSize and an IsFinite check.

A sidenote: it would be possible to compute the p-parts first, and then
take the center. However, for big permutation groups it is faster to
compute the SylowSystem of the Center of the group, than the other way around.
For polycyclic presentations there does not seem to be any difference, though.