Prevent a diff in trans.ts caused when GAP is 32-bit
The function HASH_FUNC_FOR_TRANS uses HASHKEY_BAG which uses a different
method depending on whether GAP is built as 32-bit or 64-bit. In other
words, the hash value of a transformation can be different in 32-bit
than in 64-bit, so the actual value shouldn't be displayed in the test.