Fix Mersenne Twister on 64 bit system
On 64bit systems, we would sometimes pull more random bits from
the Mersenne Twister than were strictly needed. For example, when
asked for 200 random bits, on a 32bit system we generated 7*32 = 224
bits of randomness, and discarded 24 of these. On a 64 bit system,
we generated 4*64 = 256 bits of randomness, and discard 56 of them.

This commit fixes it, and also unifies the 32bit and 64bit code paths,
resulting in somewhat easier to understand code.

Fixes #423