Add another ViewString method for inverse semigroups
This commit introduces a method for ViewString for semigroups satisfying
IsInverseSemigroup and HasGeneratorsOfSemigroup.  The rank of the filter
IsSemigroup and HasGeneratorsOfSemigroup is 14, whereas the rank of
IsInverseSemigroup is 21. Previously, if a semigroup learned that it was
inverse but did not have GeneratorsOfInverseSemigroup, then the default
ViewString method for IsInverseSemigroup was used. With this commit the
more elaborate _ViewStringForSemigroups is used instead.